some PPO examples
