+++ RockRL/tensorflow/examples/ppo/BipedalWalker-v3/BipedalWalker-v3.py
@@ -8,7 +8,6 @@ import tensorflow as tf
 for gpu in tf.config.experimental.list_physical_devices('GPU'):
     tf.config.experimental.set_memory_growth(gpu, True)
 from keras.models import Model, load_model
-# from keras.layers import Input, Dense
 from keras import layers
 
 from RockRL.utils.vectorizedEnv import VectorizedEnv, CustomEnv
@@ -20,23 +19,9 @@ from RockRL.tensorflow import PPOAgent
 
 def actor_model(input_shape, action_space):
     X_input = layers.Input(input_shape)
-    # X = Dense(512, activation='relu')(X_input)
-    # X = layers.Bidirectional(layers.LSTM(32, return_sequences=True))(X_input)
-    X = layers.LSTM(64)(X_input)
-    X = layers.Flatten()(X)
-    X = layers.Dense(32, activation='relu')(X)
-
-    # X = layers.Flatten()(X_input)
-    # X = Dense(512, activation='relu')(X)
-    # X = Dense(256, activation='relu')(X)
-    # X = Dense(64, activation='relu')(X)
-
-
-    # X = layers.Conv1D(64, 3)(X_input)
-    # X = layers.BatchNormalization()(X)
-    # X = layers.LeakyReLU()(X)
-    # X = layers.Flatten()(X)
-    # X = layers.Dense(64, activation='relu')(X)
+    X = layers.Bidirectional(layers.LSTM(32, return_sequences=True))(X_input)
+    X = layers.LSTM(32)(X)
+    X = layers.Dense(16, activation='relu')(X)
 
     action = layers.Dense(action_space, activation="tanh")(X)
     sigma = layers.Dense(action_space, activation='softplus')(X)
@@ -47,20 +32,10 @@ def actor_model(input_shape, action_space):
 
 def critic_model(input_shape):
     X_input = layers.Input(input_shape)
-    # X = layers.Bidirectional(layers.LSTM(32, return_sequences=True))(X_input)
+    X = layers.Bidirectional(layers.LSTM(32, return_sequences=True))(X_input)
     X = layers.LSTM(32)(X_input)
-    X = layers.Flatten()(X)
-    X = layers.Dense(32, activation='relu')(X)
-    # X = Dense(512, activation='relu')(X_input)
-    # X = layers.Flatten()(X_input)
-    # X = Dense(512, activation='relu')(X)
-    # X = Dense(256, activation="relu")(X)
-    # X = Dense(64, activation="relu")(X)
-    # X = layers.Conv1D(32, 3)(X_input)
-    # X = layers.BatchNormalization()(X)
-    # X = layers.LeakyReLU()(X)
-    # X = layers.Flatten()(X)
-    # X = layers.Dense(64, activation='relu')(X)
+    X = layers.Dense(16, activation='relu')(X)
+
     value = layers.Dense(1, activation=None)(X)
 
     model = Model(inputs = X_input, outputs = value)
@@ -70,26 +45,26 @@ def critic_model(input_shape):
 if __name__ == "__main__":
     env_name = 'BipedalWalker-v3'
 
-    num_envs = 24
-    env = VectorizedEnv(env_object=CustomEnv, custom_env_object=gym.make, os_hist_steps=16, num_envs=num_envs, id=env_name, hardcore=True)# , render_mode="human")
-    # env = CustomEnv(custom_env_object=gym.make, os_hist_steps=4, id=env_name)
+    num_envs = 48
+    env = VectorizedEnv(env_object=CustomEnv, custom_env_object=gym.make, os_hist_steps=4, num_envs=num_envs, id=env_name, hardcore=True)# , render_mode="human")
     # env = VectorizedEnv(env_object=gym.make, num_envs=num_envs, id=env_name) # , render_mode="human")
     action_space = env.action_space.shape[0]
     input_shape = env.observation_space.shape
 
     agent = PPOAgent(
-        actor = load_model("runs/1683898155/BipedalWalker-v3_actor.h5"),
-        critic = load_model("runs/1683898155/BipedalWalker-v3_critic.h5"),
+        actor = load_model("runs/1684126402/BipedalWalker-v3_actor.h5"),
+        critic = load_model("runs/1684126402/BipedalWalker-v3_critic.h5"),
         # actor = actor_model(input_shape, action_space),
         # critic = critic_model(input_shape),
-        actor_optimizer=tf.keras.optimizers.Adam(learning_rate=0.001),
-        critic_optimizer=tf.keras.optimizers.Adam(learning_rate=0.001),
+        actor_optimizer=tf.keras.optimizers.Adam(learning_rate=0.0003),
+        critic_optimizer=tf.keras.optimizers.Adam(learning_rate=0.0003),
         action_space="continuous",
         batch_size=256,
         train_epochs=10,
         gamma=0.99,
         lamda=0.90,
         c2=0.001,
+        compile=True,
     )
     agent.actor.summary()
 
@@ -97,12 +72,12 @@ if __name__ == "__main__":
     meanAverage = MeanAverage(
         window_size=100,
         best_mean_score=-np.inf,
-        best_mean_score_episode=2000,
+        best_mean_score_episode=200,
     )
     states = env.reset()
-    reduce_lr_episode = 2000
+    reduce_lr_episode = 200
     wait_best_window = 100
-    episodes = 50000
+    episodes = 100000
     episode = 0
     while True:
 
@@ -134,13 +109,8 @@ if __name__ == "__main__":
                     reduce_lr_episode = meanAverage.best_mean_score_episode + wait_best_window
 
             if reduce_lr_episode < episode:
-                agent.reduce_learning_rate(ratio=0.98, verbose=True, min_lr = 1e-06)
+                agent.reduce_learning_rate(ratio=0.99, verbose=True, min_lr = 5e-06)
                 reduce_lr_episode = episode + wait_best_window
-
-            # if not meanAverage.is_improoving(episode):
-            # if meanAverage.best_mean_score_episode
-            #     # pass
-            #     agent.reduce_learning_rate(ratio=0.98, verbose=True, min_lr = 1e-06)
 
             print(episode, score, mean, len(_rewards), meanAverage.best_mean_score, meanAverage.best_mean_score_episode)
